% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigo_correlation.R
\name{ggplot_correlation_cliar}
\alias{ggplot_correlation_cliar}
\title{Correlations for PIGO}
\usage{
ggplot_correlation_cliar(data, x, y, filename = NULL)
}
\arguments{
\item{data}{A data frame}

\item{x}{X-axis variable. Has to be a character vector.}

\item{y}{Y-axis variable. Has to be a character vector.}

\item{filename}{Path file to save the plot, defaults to NULL.}
}
\value{
A ggplot.
}
\description{
Correlations for PIGO
}
